feat: integrate router for workout completion navigation and refactor WorkoutTimer component
